Created the Nameset class.
The nameset class stores a set of possible
forename-surname combinations. They are loaded
from an external file. Names can then be randomly
chosen from the set to create semi-unique names.
Names are assigned a probability. For example, in
a British nameset, John and Harry are more likely
to occur than Marcus or Rohit. The names are
assigned the country they are from. There may be
multiple instances of the class for different
countries. There is a chance for names to be
double-barrelled.